+++ src/geom/camera.rs
