+++ src/geom/circle.rs
