+++ src/geom/rect.rs
