+++ src/geom/shape.rs
