+++ src/lib.rs
@@ -12,9 +12,9 @@ mod main_loop;
 pub(crate) mod util;
 
 pub mod data;
+pub mod geom;
 pub mod gl;
 pub mod light;
-pub mod math;
 pub mod pass;
 pub mod plot;
 pub mod text;
@@ -32,4 +32,3 @@ pub use context::Context;
 pub use error::{FrameError, InitError};
 pub use input::{Event, InputState, Key};
 pub use main_loop::main_loop;
-pub use math::{Camera, Rect, RotatedRect, Screen};
